+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 - Fix set typeFace without custom font
 - Improve rendering performance for messages without Markdown markup
+- Fix channel.watch (data payload was not sent)
+- Fix API 23 compatiblity
+- Add Attachment Border Color attrs
+- Add Message Link Text Color attrs
 - Add MessageSendManager
 - Update CustomMessageInputView
 - Update CustomMessageInputView2
-- Add Attachment Border Color attrs
-- Add Message Link Text Color attrs
 
 ## November 28th, 2019 - 3.4.1